Move error dialogs to the mixins, and the loading modals to the mixins
